+++ projects/TreeSketch/include/TreeSketchTreeObj.h
@@ -19,20 +19,20 @@ namespace zeno
     public:
         TreeObj(
             const zeno::vec4d &start,
-            const double offset_radian_min, const double offset_radian_max,
-            const double length_min, const double length_max,
-            const double radius_min, const double radius_max,
+            const float offset_radian_min, const float offset_radian_max,
+            const float length_min, const float length_max,
+            const float radius_min, const float radius_max,
             const int turn_points_num_min, const int turn_points_num_max,
-            const double turn_points_offset_min, const double turn_points_offset_max);
+            const float turn_points_offset_min, const float turn_points_offset_max);
 
         void create_branchs(
             const int num_min, const int num_max,
-            const double offset_start_min, const double offset_start_max,
-            const double offset_radian_min, const double offset_radian_max,
-            const double length_min, const double length_max,
-            const double radius_min, const double radius_max,
+            const float offset_start_min, const float offset_start_max,
+            const float offset_radian_min, const float offset_radian_max,
+            const float length_min, const float length_max,
+            const float radius_min, const float radius_max,
             const int turn_points_num_min, const int turn_points_num_max,
-            const double turn_points_offset_min, const double turn_points_offset_max);
+            const float turn_points_offset_min, const float turn_points_offset_max);
 
         void set_leaves();
 
@@ -44,17 +44,17 @@ namespace zeno
     {
         zeno::vec4d _start;
         zeno::vec4d _direction;
-        double _length;
-        double _radius;
+        float _length;
+        float _radius;
         std::vector<zeno::vec4d> _turn_points;
         bool _hasLeaf;
         std::vector<std::unique_ptr<BranchObj>> _children;
 
         BranchObj(
-            const zeno::vec4d &start, const zeno::vec4d &direction, const double length,
-            const double radius, const std::vector<zeno::vec4d> &turn_points);
+            const zeno::vec4d &start, const zeno::vec4d &direction, const float length,
+            const float radius, const std::vector<zeno::vec4d> &turn_points);
 
-        zeno::vec4d calculate_child_start(const double offset_start_min, const double offset_start_max);
+        zeno::vec4d calculate_child_start(const float offset_start_min, const float offset_start_max);
 
     }; // BranchObj
 }
